+++ CHANGES.md
@@ -2,6 +2,10 @@
 
 ## 0.0.6 (unreleased)
 
+-   Add superevent management functionality for triggers from the `gracedb-dev1`
+    server. Primary logic to create/update superevents in the `parse_trigger`
+    function.
+
 -   Generate GCN Circular drafts using
     [``ligo-followup-advocate``](https://git.ligo.org/emfollow/ligo-followup-advocate).